requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-filter==23.5
pillow==10.1.0
python-dateutil==2.8.2
